refactor: move wildcard selector
