+++ tests/test_registry_scanner_safety.py
@@ -85,6 +85,7 @@ EXPECTED_PYTHON_PACKAGE_FILES = {
     "easyuse_anima/wildcard/mode.py",
     "easyuse_anima/wildcard/models.py",
     "easyuse_anima/wildcard/seed.py",
+    "easyuse_anima/wildcard/selector.py",
     "easyuse_anima/wildcard/snapshot.py",
     "easyuse_anima/wildcard/sources.py",
     "easyuse_anima/registration.py",
@@ -230,6 +231,7 @@ class RegistryScannerSafetyTests(unittest.TestCase):
             "easyuse_anima/prompt/data.py",
             "easyuse_anima/prompt/fields.py",
             "easyuse_anima/wildcard/mode.py",
+            "easyuse_anima/wildcard/selector.py",
             "nodes.py",
             "prompt_translation.py",
             "settings.py",
