+++ wildcard_engine.py
@@ -134,6 +134,11 @@ except ImportError:
         normalize_wildcard_mode,
     )
 
+try:
+    from .easyuse_anima.wildcard.selector import _Selector
+except ImportError:
+    from easyuse_anima.wildcard.selector import _Selector
+
 COMMENT_RE = re.compile(r"^\s*#.*(?:\n|$)", re.MULTILINE)
 DYNAMIC_RE = re.compile(r"(?<![\\%])\{((?:[^{}]|(?<=\\)[{}])*?)(?<!\\)\}")
 WILDCARD_RE = re.compile(r"__(?P<keyword>[\w.\-+/*\\]+?)__", re.IGNORECASE)
@@ -439,63 +444,6 @@ def wildcard_sources_signature(extra_paths: str | None = None, roots: Iterable[P
         roots if roots is not None else resolve_wildcard_roots(extra_paths)
     )
     return snapshot.public_signature()
-
-
-class _Selector:
-    def __init__(self, seed: int, sequential: bool):
-        self.seed = normalize_seed(seed)
-        self.sequential = sequential
-        self.rng = (
-            None
-            if self.sequential
-            else np.random.Generator(np.random.PCG64(self.seed))
-        )
-
-    def count_from_range(self, minimum: int, maximum: int) -> int:
-        minimum = max(0, minimum)
-        maximum = max(minimum, maximum)
-        if minimum == maximum:
-            return minimum
-        if self.sequential:
-            return minimum + (self.seed % (maximum - minimum + 1))
-        return int(self.rng.integers(minimum, maximum + 1))
-
-    def choose_one(self, options: Sequence[WildcardOption]) -> WildcardOption | None:
-        selected = self.choose_many(options, 1)
-        return selected[0] if selected else None
-
-    def choose_many(
-        self,
-        options: Sequence[WildcardOption],
-        count: int,
-    ) -> list[WildcardOption]:
-        if not options or count <= 0:
-            return []
-        if self.sequential:
-            count = min(count, len(options))
-            start = self.seed % len(options)
-            return [options[(start + offset) % len(options)] for offset in range(count)]
-
-        weights = [max(0.0, option.weight) for option in options]
-        positive = [
-            (option, weight)
-            for option, weight in zip(options, weights)
-            if weight > 0
-        ]
-        if positive:
-            pool = [option for option, _weight in positive]
-            pool_weights = [weight for _option, weight in positive]
-        else:
-            pool = list(options)
-            pool_weights = None
-
-        count = min(count, len(pool))
-        probabilities = None
-        if pool_weights is not None:
-            total = sum(pool_weights)
-            probabilities = [weight / total for weight in pool_weights]
-        indices = self.rng.choice(len(pool), size=count, replace=False, p=probabilities)
-        return [pool[int(index)] for index in indices]
 
 
 class _WildcardLibrary:
